apps/partner-api/requirements.txt

fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
firebase-admin==6.5.0
google-cloud-firestore==2.17.0
google-cloud-secret-manager==2.20.2
google-cloud-kms==2.25.0
requests==2.32.3
ccxt==4.4.49
cryptography==43.0.1
pydantic==2.8.2
pandas==2.2.2
numpy==1.26.4
